RavenDB-15020 : initial work on slope
