+++ src/Raven.Client/Documents/Queries/TimeSeries/TimeSeriesAggregation.cs
@@ -145,7 +145,7 @@ namespace Raven.Client.Documents.Queries.TimeSeries
     public class TimeSeriesRangeAggregation : IPostJsonDeserialization, ITimeSeriesQueryStreamEntry
     {
         public long[] Count;
-        public double[] Max, Min, Last, First, Average, Sum, Percentile;
+        public double[] Max, Min, Last, First, Average, Sum, Percentile, Slope;
         public DateTime To, From;
 
         public object Key { get; private set; }
@@ -294,6 +294,7 @@ namespace Raven.Client.Documents.Queries.TimeSeries
         Count = 5,
         Average = 6,
 
-        Percentile = 7
+        Percentile = 7,
+        Slope = 8
     }
 }
