+++ src/Raven.Server/Documents/Queries/AST/Query.cs
@@ -552,357 +552,4 @@ namespace Raven.Server.Documents.Queries.AST
             return TimeValue.FromMonths(Months);
         }
     }
-
-
-    public class TimeSeriesAggregation
-    {
-        public AggregationType Aggregation;
-
-        public bool Any => NumberOfValues > 0;
-
-        public List<double> Count => _count;
-
-        public long TotalCount;
-
-        public string Name;
-
-        public int NumberOfValues => _values.Count;
-        private readonly List<double> _values;
-
-        private readonly List<double> _count;
-
-        private readonly double _percentileFactor;
-
-        private readonly List<List<double>> _rankedValues;
-
-        public TimeSeriesAggregation(AggregationType type, string name = null)
-        {
-            Aggregation = type;
-            Name = name ?? Aggregation.ToString();
-
-            _count = new List<double>();
-            _values = new List<double>();
-        }
-
-        internal TimeSeriesAggregation(double percentile, string name = null) : this(AggregationType.Percentile, name)
-        {
-            if (percentile < 0 || percentile > 100)
-                throw new ArgumentOutOfRangeException(
-                    $"Invalid argument passed to '{nameof(AggregationType.Percentile)}' aggregation method: '{percentile}'. " +
-                    "Argument must be a number between 0 and 100");
-
-            _percentileFactor = percentile / 100;
-            _rankedValues = new List<List<double>>();
-        }
-
-        public void Init()
-        {
-            _count.Clear();
-            _values.Clear();
-
-            _rankedValues?.Clear();
-        }
-
-        public void Segment(Span<StatefulTimestampValue> values, bool isRaw)
-        {
-            if (isRaw == false)
-            {
-                SegmentOnRollup(values);
-                return;
-            }
-
-            if (_count.Count < values.Length)
-            {
-                for (int i = _count.Count; i < values.Length; i++)
-                {
-                    _count.Add(0);
-                    _values.Add(0);
-                }
-            }
-
-            for (int i = 0; i < values.Length; i++)
-            {
-                var val = values[i];
-                switch (Aggregation)
-                {
-                    case AggregationType.Min:
-                        if (_count[i] == 0)
-                            _values[i] = val.Min;
-                        else
-                            _values[i] = Math.Min(_values[i], val.Min);
-                        break;
-                    case AggregationType.Max:
-                        if (_count[i] == 0)
-                            _values[i] = val.Max;
-                        else
-                            _values[i] = Math.Max(_values[i], val.Max);
-                        break;
-                    case AggregationType.Sum:
-                    case AggregationType.Average:
-                        _values[i] = _values[i] + val.Sum;
-                        break;
-                    case AggregationType.First:
-                        if (_count[i] == 0)
-                            _values[i] = val.First;
-                        break;
-                    case AggregationType.Last:
-                        _values[i] = val.Last;
-                        break;
-                    case AggregationType.Count:
-                        break;
-                    case AggregationType.Percentile:
-                        throw new InvalidOperationException($"Cannot use method '{nameof(Segment)}' on aggregation type '{nameof(AggregationType.Percentile)}' ");
-                    default:
-                        throw new ArgumentOutOfRangeException("Unknown aggregation operation: " + Aggregation);
-                }
-
-                _count[i] += values[i].Count;
-            }
-
-            TotalCount += (long)_count[0];
-        }
-
-        public void Step(Span<double> values, bool isRaw)
-        {
-            if (isRaw == false)
-            {
-                StepOnRollup(values);
-                return;
-            }
-
-            if (_count.Count < values.Length)
-            {
-                for (int i = _count.Count; i < values.Length; i++)
-                {
-                    _count.Add(0);
-                    _values.Add(0);
-                }
-            }
-
-            for (int i = 0; i < values.Length; i++)
-            {
-                var val = values[i];
-                switch (Aggregation)
-                {
-                    case AggregationType.Min:
-                        if (_count[i] == 0)
-                            _values[i] = val;
-                        else
-                            _values[i] = Math.Min(_values[i], val);
-                        break;
-                    case AggregationType.Max:
-                        if (_count[i] == 0)
-                            _values[i] = val;
-                        else
-                            _values[i] = Math.Max(_values[i], val);
-                        break;
-                    case AggregationType.Sum:
-                    case AggregationType.Average:
-                        _values[i] = _values[i] + val;
-                        break;
-                    case AggregationType.First:
-                        if (_count[i] == 0)
-                            _values[i] = val;
-                        break;
-                    case AggregationType.Last:
-                        _values[i] = val;
-                        break;
-                    case AggregationType.Count:
-                        break;
-                    case AggregationType.Percentile:
-                        for (var j = _rankedValues.Count; j < values.Length; j++)
-                        {
-                            _rankedValues.Add(new List<double>());
-                        }
-
-                        var currentRankedList = _rankedValues[i];
-                        var loc = currentRankedList.BinarySearch(val);
-                        if (loc >= 0)
-                            currentRankedList.Insert(loc + 1, val);
-                        else
-                            currentRankedList.Insert(~loc, val);
-
-                        break;
-                    default:
-                        throw new ArgumentOutOfRangeException("Unknown aggregation operation: " + Aggregation);
-                }
-
-                _count[i]++;
-            }
-
-            TotalCount++;
-        }
-
-        private void SegmentOnRollup(Span<StatefulTimestampValue> values)
-        {
-            Debug.Assert(values.Length % 6 == 0);
-            var originalNumOfValues = values.Length / 6;
-            if (_count.Count < originalNumOfValues)
-            {
-                for (int i = _count.Count; i < originalNumOfValues; i++)
-                {
-                    _count.Add(0L);
-                    _values.Add(0d);
-                }
-            }
-
-            for (int i = 0; i < originalNumOfValues; i++)
-            {
-                var index = i * 6;
-                var val = Aggregation == AggregationType.Average ? values[index + (int)AggregationType.Sum] : values[index + (int)Aggregation];
-
-                switch (Aggregation)
-                {
-                    case AggregationType.Min:
-                        if (_count[i] == 0)
-                            _values[i] = val.Min;
-                        else
-                            _values[i] = Math.Min(_values[i], val.Min);
-                        break;
-                    case AggregationType.Max:
-                        if (_count[i] == 0)
-                            _values[i] = val.Max;
-                        else
-                            _values[i] = Math.Max(_values[i], val.Max);
-                        break;
-                    case AggregationType.Average:
-                    case AggregationType.Sum:
-                        _values[i] = _values[i] + val.Sum;
-                        break;
-                    case AggregationType.First:
-                        if (_count[i] == 0)
-                            _values[i] = val.First;
-                        break;
-                    case AggregationType.Last:
-                        _values[i] = val.Last;
-                        break;
-                    case AggregationType.Count:
-                        break;
-                    case AggregationType.Percentile:
-                        throw new InvalidOperationException($"Cannot use aggregation method '{nameof(AggregationType.Percentile)}' on rolled-up time series");
-                    default:
-                        throw new ArgumentOutOfRangeException("Unknown aggregation operation: " + Aggregation);
-                }
-
-                val = values[index + (int)AggregationType.Count];
-                _count[i] += (long)val.Sum;
-            }
-
-            TotalCount += (long)_count[0];
-        }
-
-        private void StepOnRollup(Span<double> values)
-        {
-            Debug.Assert(values.Length % 6 == 0);
-            var originalNumOfValues = values.Length / 6;
-            if (_count.Count < originalNumOfValues)
-            {
-                for (int i = _count.Count; i < originalNumOfValues; i++)
-                {
-                    _count.Add(0L);
-                    _values.Add(0d);
-                }
-            }
-
-            for (int i = 0; i < originalNumOfValues; i++)
-            {
-                var index = i * 6;
-                double val;
-                switch (Aggregation)
-                {
-                    case AggregationType.Min:
-                        val = values[index + (int)AggregationType.Min];
-                        if (_count[i] == 0)
-                            _values[i] = val;
-                        else
-                            _values[i] = Math.Min(_values[i], val);
-                        break;
-                    case AggregationType.Max:
-                        val = values[index + (int)AggregationType.Max];
-                        if (_count[i] == 0)
-                            _values[i] = val;
-                        else
-                            _values[i] = Math.Max(_values[i], val);
-                        break;
-                    case AggregationType.Sum:
-                    case AggregationType.Average:
-                        val = values[index + (int)AggregationType.Sum];
-                        _values[i] = _values[i] + val;
-                        break;
-                    case AggregationType.First:
-                        val = values[index + (int)AggregationType.First];
-                        if (_count[i] == 0)
-                            _values[i] = val;
-                        break;
-                    case AggregationType.Last:
-                        val = values[index + (int)AggregationType.Last];
-                        _values[i] = val;
-                        break;
-                    case AggregationType.Count:
-                        break;
-                    case AggregationType.Percentile:
-                        throw new InvalidOperationException($"Cannot use aggregation method '{nameof(AggregationType.Percentile)}' on rolled-up time series");
-                    default:
-                        throw new ArgumentOutOfRangeException("Unknown aggregation operation: " + Aggregation);
-                }
-
-                _count[i] += (long)values[index + (int)AggregationType.Count];
-            }
-
-            TotalCount += (long)values[(int)AggregationType.Count];
-        }
-
-        public IEnumerable<double> GetFinalValues(double? scale = null)
-        {
-            switch (Aggregation)
-            {
-                case AggregationType.Min:
-                case AggregationType.Max:
-                case AggregationType.First:
-                case AggregationType.Last:
-                case AggregationType.Sum:
-                    break;
-                case AggregationType.Count:
-                    return _count;
-                case AggregationType.Average:
-                    for (int i = 0; i < _values.Count; i++)
-                    {
-                        if (_count[i] == 0)
-                        {
-                            _values[i] = double.NaN;
-                            continue;
-                        }
-
-                        _values[i] = _values[i] / _count[i];
-                    }
-
-                    break;
-                case AggregationType.Percentile:
-                    return GetPercentiles();
-                default:
-                    throw new ArgumentOutOfRangeException("Unknown aggregation operation: " + Aggregation);
-            }
-
-            if (scale.HasValue)
-                return _values.Select(x => x * scale.Value);
-
-            return _values;
-        }
-
-        private IEnumerable<double> GetPercentiles()
-        {
-            /*if (scale.HasValue)
-            {
-                // todo
-                return _values.Select(x => x * scale.Value);
-            }*/
-
-            foreach (var list in _rankedValues)
-            {
-                var index = (int)Math.Ceiling(_percentileFactor * list.Count);
-
-                yield return list[index - 1];
-            }
-        }
-    }
 }
